+++ application/views/receiving-edit_view.php
@@ -1,9 +1,9 @@
 <h2>{action} Receiving </h2>
 {error_messages}
 <form action="/Receiving/save" method="post" enctype="multipart/form-data">
-    {id}
-    {receiving}
-    {price}
-    {submit}
+    {fid}
+    {freceiving}
+    {fprice}
+    {zsubmit}
     <a class="btn btn-default" role="button" href="/Receiving/cancel">Cancel</a>
 </form>